Separate view file for bulk edit views
