+++ src/js/views/patients/worklist/worklist-list.scss
@@ -110,12 +110,12 @@
   }
 }
 
-.worklist-list__multi-edit-select {
+.worklist-list__bulk-edit-select {
   margin-right: 16px;
   vertical-align: middle;
 }
 
-.worklist-list__multi-edit-cancel {
+.worklist-list__bulk-edit-cancel {
   color: $black-60;
   cursor: pointer;
   font-size: 14px;
